Don't calculate financial years; pass them in.
Because there may be exceptions, e.g. a longer first financial year,
and that complexity doesn't belong in this library.

This is a backwards incompatible change. Sorry about that! You can
trivially migrate your code by copying the removed FinancialYears code
to your own app and using it to generate the list of years.